Adding Custody Test

+++ src/NewOrder/NewOrder.Custody/Domain/ICustodyDatabase.cs
@@ -2,7 +2,7 @@
 {
     public interface ICustodyDatabase
     {
-        public Custody Get(int accountNumber);
+        public Custody Get(long accountNumber);
         public void Save(Custody custody);
     }
 }
